mes-slack-bot/requirements.txt
pymssql
requests
python-dotenv
streamlit
plotly
pandas
